WoT restructure: move overlay-only scroll collaborators into overlay/scroll/
scrollSync.ts + nativeScrollInput.ts are imported only by Overlays, so they belong inside the
overlay slice, not the shared top-level scroll/ (which keeps scroll.ts, the viewport Scroll module
used via the facade). Grouped in overlay/scroll/ rather than flat in overlay/ so the slice stays
navigable as more collaborators extract. Import paths only. Gated green: test:types + eslint +
test:walkontable 749/0/2.
